// src/api/generated/iMSIncidentManagementSystem.schemas.ts
export interface CategoryCreate {
  name: string;
  description?: string | null;
}

export interface CategoryUpdate {
  name?: string | null;
  description?: string | null;
}